add dialogs add shared module change styles

+++ src/app/components/todo-item/todo-item.component.ts
@@ -1,5 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { Todo, TodoService } from '../../services/todo.service';
+import { Dialog } from '@angular/cdk/dialog';
+import { ConfirmDialogComponent } from '../../shared/components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-todo-item',
@@ -9,11 +11,20 @@ import { Todo, TodoService } from '../../services/todo.service';
 export class TodoItemComponent {
   @Input() todo!: Todo;
 
-  constructor(private todoService: TodoService) {
+  constructor(private todoService: TodoService, private dialog: Dialog) {
   }
 
   deleteTodo() {
-    this.todoService.removeTodo(this.todo);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        message: 'Do you really want to remove this todo?'
+      }
+    });
+    dialogRef.closed.subscribe({
+      next: (res) => {
+        if (res) this.todoService.removeTodo(this.todo);
+      }
+    });
   }
 
   markForEdit() {
@@ -21,7 +32,8 @@ export class TodoItemComponent {
     this.todoService.editTodo(this.todo);
   }
 
-  checkTodo() {
+  checkTodo(event: Event) {
+    if (event.target !== event.currentTarget) return;
     this.todo.checked = !this.todo.checked;
     this.todoService.editTodo(this.todo);
   }
